% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_meetings.R
\name{group_plan}
\alias{group_plan}
\title{Group meeting planner}
\usage{
group_plan(ndates, namelist, nrep = 2)
}
\arguments{
\item{ndates}{Number of days in the plan}

\item{namelist}{Tibble containing columns name and journalclub}

\item{nrep}{Max number of journal clubs per person}
}
\description{
Generates random group meeting plans until the following conditions are met:
No people having presentations in consequitive weeks or on the same day.
All project presentations unique.
No more than two journal clubs per person
}
\examples{

library(tidyverse)
nms <- tibble::tribble(
~name, ~journalclub,
"Asger", "Electrochemistry",
"Emil", "Electrochemistry",
"Magnus", "Electrochemistry",
"Matteo", "Electrochemistry",
"Andreas", "Polymer chemistry",
"Paolo", "Polymer chemistry",
"Stefan", "Polymer chemistry",
"Monica", "Electrochemistry",
"Siad", "Electrochemistry",
"Christina", "Polymer chemistry",
"Annika", "Polymer chemistry",
"Cecilie", "Polymer chemistry",
"Mahdi", "Polymer chemistry",
"Surnimal", "Polymer chemistry",
"Xinming", "Electrochemistry",
"Lizhi", "Polymer chemistry",
"Mie", "Polymer chemistry"
#"Kim", "Polymer chemistry",
)

plan <- group_plan(13, nms, nrep = 2)

readr::write_excel_csv(plan, "2017_fall_group_meetings.csv")

nms \%>\% dplyr::filter(journalclub=="Electrochemistry") \%>\% dplyr::pull(name) \%>\% paste(collapse = ", ")
nms \%>\% dplyr::filter(journalclub=="Polymer chemistry") \%>\% dplyr::pull(name) \%>\% paste(collapse = ", ")
}
