% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_journal}
\alias{new_journal}
\title{New journal}
\usage{
new_journal(title, date = gsub("-", "", Sys.Date()), open = interactive(),
  file = NULL)
}
\arguments{
\item{title}{Title of the new journal}

\item{date}{Date given as "YYYYMMDD" e.g. "20170901"}

\item{open}{Optional. Whether to open the file after creation.}

\item{file}{Optional. Filename is generated from date and title if not given.}
}
\description{
New journal
}
